begin to develop file structure
