+++ lib/foods/eventHandler.js
@@ -1,0 +1,7 @@
+// const deleteFood = require('./foods/requests')
+// const editFood = require('./foods/requests')
+
+// $('.foods').on('click', '.delete-food-btn', deleteFood);
+// $('.foods').on('blur', '.food', editFood);
+// $('#submit').on('click', addNewFood);
+// $('#foodFilter').on('keyup', filterFoods);
