+++ lib/index.js
@@ -1,136 +1,144 @@
 // Foods
 
-const getFoods = () => {
-  return fetch(`https://evening-peak-25213.herokuapp.com/api/v1/foods`)
-  .then(response => handleResponse(response))
-  .then(foods => getAllFoods(foods))
-  .catch(error => console.log({error}));
-}
+const getFoods = require('./foods/requests')
+const deleteFood = require('./foods/requests')
+const editFood = require('./foods/requests')
+// const 
 
-const handleResponse = (response) => {
-  return response.json()
-  .then(json => {
-    if(!response.ok){
-      const error = {
-        status: response.status,
-        statusText: response.statusText,
-        json,
-      };
-      return Promise.reject(error);
-    }
-  return json;
-  })
-}
-
-const getAllFoods = (foods) => {
-  return foods.forEach((food) => {
-    prependFood(food)
-  }
-)}
-
-const prependFood = (food) => {
-  $('.foods').prepend(`
-    <article class='food' data-food-id=${food.id}>
-      <div contenteditable="true">
-        <p class='name'> ${food.name} </p>
-      </div>
-      <div contenteditable="true">
-        <p class='calories' >${food.calories} </p>
-      </div>
-        <p class='delete'><button data-food-id=${food.id} class='delete-food-btn'><i class="fa fa-trash"></i></button></p>
-    </article>`)
-}
-
-const deleteFood = (event) => {
-  event.preventDefault();
-  var foodId = $('.delete-food-btn').data('food-id').toString();
-  return fetch(`https://evening-peak-25213.herokuapp.com/api/v1/foods` + '/' + foodId.toString(), {
-    method: 'DELETE',
-    headers:
-     {'Content-Type': 'application/json'}
-  })
-  .then(getFoods)
-  .catch(error => console.log({error}));
-}
-
-function editFood()  {
-  var foodId = $(this).data('food-id').toString();
-  var html = $(this).html();
-  var updatedName = (this.getElementsByClassName('name')[0]).innerText;
-  var updatedCalories = (this.getElementsByClassName('calories')[0]).innerText;
-
-
-  return fetch(`https://evening-peak-25213.herokuapp.com/api/v1/foods` + '/' + foodId, {
-    method: 'PUT',
-    headers:
-      {'Content-Type': 'application/json'},
-    body: JSON.stringify({food: {name: updatedName, calories: updatedCalories}})
-  })
-  .catch(error => console.log({error}));
-}
-
-const addNewFood = () => {
-  event.preventDefault()
-  $('#new-name-error').hide()
-  $('#new-calories-error').hide()
-  if ($('#new-name').val() === "") {
-    $('#new-name-error').css('display', 'block')
-    if ($('#new-calories').val() === "") {
-      $('#new-calories-error').css('display', 'block')
-    }
-  }
-  else if ($('#new-calories').val() === "") {
-    $('#new-calories-error').css('display', 'block')
-  }
-  else {
-    var data = { name: $('#new-name').val(), calories: $('#new-calories').val()}
-    return fetch(`https://evening-peak-25213.herokuapp.com/api/v1/foods`, {
-      method: 'post',
-      headers:
-      { 'Content-Type': 'application/json' },
-      body: JSON.stringify(data) })
-      .then(response => response.json())
-      .then(newFood => prependFood(newFood))
-      .then(resetFoodForm)
-      .catch(error => console.log({error}))
-  }
-};
-
-const resetFoodForm = () => {
-  $(".new-food")[0].reset();
-}
-
-function filterFoods() {
-  var input, filter, foods, food, name, i;
-  input = document.getElementById('foodFilter');
-  filter = input.value.toUpperCase();
-  foods = document.getElementsByClassName("foods");
-  food = document.getElementsByClassName('food');
-
-  for (i = 0; i < food.length; i++) {
-    name = food[i].getElementsByClassName("name")[0];
-    if (name.innerHTML.toUpperCase().indexOf(filter) > -1) {
-      food[i].style.display = "";
-    } else {
-      food[i].style.display = "none";
-    }
-  }
-}
 
 $(document).ready(() => {
   getFoods();
+  $('.foods').on('click', '.delete-food-btn', deleteFood);
+  $('.foods').on('blur', '.food', editFood);
+  // $('#submit').on('click', addNewFood);
+  // $('#foodFilter').on('keyup', filterFoods);
 });
 
-$('.foods').on('click', '.delete-food-btn', deleteFood);
-$('.foods').on('blur', '.food', editFood);
-$('#submit').on('click', addNewFood);
-$('#foodFilter').on('keyup', filterFoods);
+// const getFoods = () => {
+//   return fetch(`https://evening-peak-25213.herokuapp.com/api/v1/foods`)
+//   .then(response => handleResponse(response))
+//   .then(foods => getAllFoods(foods))
+//   .catch(error => console.log({error}));
+// }
+
+// const handleResponse = (response) => {
+//   return response.json()
+//   .then(json => {
+//     if(!response.ok){
+//       const error = {
+//         status: response.status,
+//         statusText: response.statusText,
+//         json,
+//       };
+//       return Promise.reject(error);
+//     }
+//   return json;
+//   })
+// }
+
+// const getAllFoods = (foods) => {
+//   return foods.forEach((food) => {
+//     prependFood(food)
+//   }
+// )}
+
+// const prependFood = (food) => {
+//   $('.foods').prepend(`
+//     <article class='food' data-food-id=${food.id}>
+//       <div contenteditable="true">
+//         <p class='name'> ${food.name} </p>
+//       </div>
+//       <div contenteditable="true">
+//         <p class='calories' >${food.calories} </p>
+//       </div>
+//         <p class='delete'><button data-food-id=${food.id} class='delete-food-btn'><i class="fa fa-trash"></i></button></p>
+//     </article>`)
+// }
+
+// const deleteFood = (event) => {
+//   event.preventDefault();
+//   var foodId = $('.delete-food-btn').data('food-id').toString();
+//   return fetch(`https://evening-peak-25213.herokuapp.com/api/v1/foods` + '/' + foodId.toString(), {
+//     method: 'DELETE',
+//     headers:
+//      {'Content-Type': 'application/json'}
+//   })
+//   .then(getFoods)
+//   .catch(error => console.log({error}));
+// }
+
+// function editFood()  {
+//   var foodId = $(this).data('food-id').toString();
+//   var html = $(this).html();
+//   var updatedName = (this.getElementsByClassName('name')[0]).innerText;
+//   var updatedCalories = (this.getElementsByClassName('calories')[0]).innerText;
 
 
-module.exports = {addNewFood}
+//   return fetch(`https://evening-peak-25213.herokuapp.com/api/v1/foods` + '/' + foodId, {
+//     method: 'PUT',
+//     headers:
+//       {'Content-Type': 'application/json'},
+//     body: JSON.stringify({food: {name: updatedName, calories: updatedCalories}})
+//   })
+//   .catch(error => console.log({error}));
+// }
 
+// const addNewFood = () => {
+//   event.preventDefault()
+//   $('#new-name-error').hide()
+//   $('#new-calories-error').hide()
+//   if ($('#new-name').val() === "") {
+//     $('#new-name-error').css('display', 'block')
+//     if ($('#new-calories').val() === "") {
+//       $('#new-calories-error').css('display', 'block')
+//     }
+//   }
+//   else if ($('#new-calories').val() === "") {
+//     $('#new-calories-error').css('display', 'block')
+//   }
+//   else {
+//     var data = { name: $('#new-name').val(), calories: $('#new-calories').val()}
+//     return fetch(`https://evening-peak-25213.herokuapp.com/api/v1/foods`, {
+//       method: 'post',
+//       headers:
+//       { 'Content-Type': 'application/json' },
+//       body: JSON.stringify(data) })
+//       .then(response => response.json())
+//       .then(newFood => prependFood(newFood))
+//       .then(resetFoodForm)
+//       .catch(error => console.log({error}))
+//   }
+// };
 
+// const resetFoodForm = () => {
+//   $(".new-food")[0].reset();
+// }
 
+// function filterFoods() {
+//   var input, filter, foods, food, name, i;
+//   input = document.getElementById('foodFilter');
+//   filter = input.value.toUpperCase();
+//   foods = document.getElementsByClassName("foods");
+//   food = document.getElementsByClassName('food');
+
+//   for (i = 0; i < food.length; i++) {
+//     name = food[i].getElementsByClassName("name")[0];
+//     if (name.innerHTML.toUpperCase().indexOf(filter) > -1) {
+//       food[i].style.display = "";
+//     } else {
+//       food[i].style.display = "none";
+//     }
+//   }
+// }
+
+// $(document).ready(() => {
+//   getFoods();
+// });
+
+// $('.foods').on('click', '.delete-food-btn', deleteFood);
+// $('.foods').on('blur', '.food', editFood);
+// $('#submit').on('click', addNewFood);
+// $('#foodFilter').on('keyup', filterFoods);
 
 
 // Meals
@@ -548,12 +556,12 @@ const postDeleteTotalRemaining = (food, meal, currentTotalRemaining) => {
   return (currentTotalRemaining + calories)
 }
 
-$('.new-food-btn').on('click', redirectNewFood)
-$('#diaryFilter').on('keyup', filterDiary);
-$('.add-meal-btn').on('click', getCheckBoxIds);
-$('.breakfasts').on('click', '.delete-meal-btn', deleteBreakfastMeal);
-$('.snacks').on('click', '.delete-meal-btn', deleteSnackMeal);
-$('.lunches').on('click', '.delete-meal-btn', deleteLunchMeal);
-$('.dinners').on('click', '.delete-meal-btn', deleteDinnerMeal);
+// $('.new-food-btn').on('click', redirectNewFood)
+// $('#diaryFilter').on('keyup', filterDiary);
+// $('.add-meal-btn').on('click', getCheckBoxIds);
+// $('.breakfasts').on('click', '.delete-meal-btn', deleteBreakfastMeal);
+// $('.snacks').on('click', '.delete-meal-btn', deleteSnackMeal);
+// $('.lunches').on('click', '.delete-meal-btn', deleteLunchMeal);
+// $('.dinners').on('click', '.delete-meal-btn', deleteDinnerMeal);
